// src/lib/types.ts
interface CalendarEvent {
  event_id: number | string;
  title: React.ReactNode;
  subtitle?: React.ReactNode;
  start: Date;
  end: Date;
  recurring?: RRule;
  disabled?: boolean;
  color?: string;
  textColor?: string;
  editable?: boolean;
  deletable?: boolean;
  draggable?: boolean;
  allDay?: boolean;
  /**
   * @default " "
   * passed as a children to mui <Avatar /> component
   */
  agendaAvatar?: React.ReactElement | string;
}
// ...
export type ProcessedEvent = CalendarEvent & Record<string, any>;